pyproject.toml
[project]
name = "mo-april"
version = "0.1.0"
description = "Learning platform: blog, syllabus, topics, coding, Q&A, references — modular monolith (split-ready)."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<0.116.0",
    "uvicorn[standard]>=0.32.0,<0.33.0",
    "gunicorn>=23.0.0,<24.0.0",
    "sqlmodel>=0.0.22,<0.0.30",
    "sqlalchemy>=2.0.36,<3.0.0",
    "greenlet>=3.0.0,<4.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
    "asyncpg>=0.30.0,<0.31.0",
    "aiomysql>=0.2.0,<0.3.0",
    "pymysql>=1.1.0,<2.0.0",
    "pydantic-settings>=2.6.0,<3.0.0",
    "pydantic[email]>=2.0,<3.0",
    "python-multipart>=0.0.12,<0.1.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "slowapi>=0.1.9,<0.2.0",
    "jinja2>=3.1.0,<4.0.0",
    "rich>=13.7.0,<14.0.0",
]

[tool.uv]
package = false
